Pass feature specs to online retriever
