+++ storage/api/src/main/java/feast/storage/common/testing/TestUtil.java
@@ -1,0 +1,187 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright 2018-2019 The Feast Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package feast.storage.common.testing;
+
+import com.google.protobuf.ByteString;
+import com.google.protobuf.util.Timestamps;
+import feast.core.FeatureSetProto.FeatureSet;
+import feast.core.FeatureSetProto.FeatureSetSpec;
+import feast.types.FeatureRowProto.FeatureRow;
+import feast.types.FeatureRowProto.FeatureRow.Builder;
+import feast.types.FieldProto.Field;
+import feast.types.ValueProto.*;
+import java.util.concurrent.ThreadLocalRandom;
+import org.apache.commons.lang3.RandomStringUtils;
+
+@SuppressWarnings("WeakerAccess")
+public class TestUtil {
+
+  /**
+   * Create a Feature Row with random value according to the FeatureSetSpec
+   *
+   * <p>See {@link #createRandomFeatureRow(FeatureSet, int)}
+   */
+  public static FeatureRow createRandomFeatureRow(FeatureSet featureSet) {
+    ThreadLocalRandom random = ThreadLocalRandom.current();
+    int randomStringSizeMaxSize = 12;
+    return createRandomFeatureRow(featureSet, random.nextInt(0, randomStringSizeMaxSize) + 4);
+  }
+
+  /**
+   * Create a Feature Row with random value according to the FeatureSet.
+   *
+   * <p>The Feature Row created contains fields according to the entities and features defined in
+   * FeatureSet, matching the value type of the field, with randomized value for testing.
+   *
+   * @param featureSet {@link FeatureSet}
+   * @param randomStringSize number of characters for the generated random string
+   * @return {@link FeatureRow}
+   */
+  public static FeatureRow createRandomFeatureRow(FeatureSet featureSet, int randomStringSize) {
+    Builder builder =
+        FeatureRow.newBuilder()
+            .setFeatureSet(getFeatureSetReference(featureSet))
+            .setEventTimestamp(Timestamps.fromMillis(System.currentTimeMillis()));
+
+    featureSet
+        .getSpec()
+        .getEntitiesList()
+        .forEach(
+            field -> {
+              builder.addFields(
+                  Field.newBuilder()
+                      .setName(field.getName())
+                      .setValue(createRandomValue(field.getValueType(), randomStringSize))
+                      .build());
+            });
+
+    featureSet
+        .getSpec()
+        .getFeaturesList()
+        .forEach(
+            field -> {
+              builder.addFields(
+                  Field.newBuilder()
+                      .setName(field.getName())
+                      .setValue(createRandomValue(field.getValueType(), randomStringSize))
+                      .build());
+            });
+
+    return builder.build();
+  }
+
+  private static String getFeatureSetReference(FeatureSet featureSet) {
+    FeatureSetSpec spec = featureSet.getSpec();
+    return String.format("%s/%s:%d", spec.getProject(), spec.getName(), spec.getVersion());
+  }
+
+  /**
+   * Create a random Feast {@link Value} of {@link ValueType.Enum}.
+   *
+   * @param type {@link ValueType.Enum}
+   * @param randomStringSize number of characters for the generated random string
+   * @return {@link Value}
+   */
+  public static Value createRandomValue(ValueType.Enum type, int randomStringSize) {
+    Value.Builder builder = Value.newBuilder();
+    ThreadLocalRandom random = ThreadLocalRandom.current();
+
+    switch (type) {
+      case INVALID:
+      case UNRECOGNIZED:
+        throw new IllegalArgumentException("Invalid ValueType: " + type);
+      case BYTES:
+        builder.setBytesVal(
+            ByteString.copyFrom(RandomStringUtils.randomAlphanumeric(randomStringSize).getBytes()));
+        break;
+      case STRING:
+        builder.setStringVal(RandomStringUtils.randomAlphanumeric(randomStringSize));
+        break;
+      case INT32:
+        builder.setInt32Val(random.nextInt());
+        break;
+      case INT64:
+        builder.setInt64Val(random.nextLong());
+        break;
+      case DOUBLE:
+        builder.setDoubleVal(random.nextDouble());
+        break;
+      case FLOAT:
+        builder.setFloatVal(random.nextFloat());
+        break;
+      case BOOL:
+        builder.setBoolVal(random.nextBoolean());
+        break;
+      case BYTES_LIST:
+        builder.setBytesListVal(
+            BytesList.newBuilder()
+                .addVal(
+                    ByteString.copyFrom(
+                        RandomStringUtils.randomAlphanumeric(randomStringSize).getBytes()))
+                .build());
+        break;
+      case STRING_LIST:
+        builder.setStringListVal(
+            StringList.newBuilder()
+                .addVal(RandomStringUtils.randomAlphanumeric(randomStringSize))
+                .build());
+        break;
+      case INT32_LIST:
+        builder.setInt32ListVal(Int32List.newBuilder().addVal(random.nextInt()).build());
+        break;
+      case INT64_LIST:
+        builder.setInt64ListVal(Int64List.newBuilder().addVal(random.nextLong()).build());
+        break;
+      case DOUBLE_LIST:
+        builder.setDoubleListVal(DoubleList.newBuilder().addVal(random.nextDouble()).build());
+        break;
+      case FLOAT_LIST:
+        builder.setFloatListVal(FloatList.newBuilder().addVal(random.nextFloat()).build());
+        break;
+      case BOOL_LIST:
+        builder.setBoolListVal(BoolList.newBuilder().addVal(random.nextBoolean()).build());
+        break;
+    }
+    return builder.build();
+  }
+
+  /**
+   * Create a field object with given name and type.
+   *
+   * @param name of the field.
+   * @param value of the field. Should be compatible with the valuetype given.
+   * @param valueType type of the field.
+   * @return Field object
+   */
+  public static Field field(String name, Object value, ValueType.Enum valueType) {
+    Field.Builder fieldBuilder = Field.newBuilder().setName(name);
+    switch (valueType) {
+      case INT32:
+        return fieldBuilder.setValue(Value.newBuilder().setInt32Val((int) value)).build();
+      case INT64:
+        return fieldBuilder.setValue(Value.newBuilder().setInt64Val((int) value)).build();
+      case FLOAT:
+        return fieldBuilder.setValue(Value.newBuilder().setFloatVal((float) value)).build();
+      case DOUBLE:
+        return fieldBuilder.setValue(Value.newBuilder().setDoubleVal((double) value)).build();
+      case STRING:
+        return fieldBuilder.setValue(Value.newBuilder().setStringVal((String) value)).build();
+      default:
+        throw new IllegalStateException("Unexpected valueType: " + value.getClass());
+    }
+  }
+}
